pyproject.toml
[tool.poetry]
name = "dojo"
version = "0.1.0"
description = ""
authors = ["Milan Boháček <244656@muni.cz>"]
readme = "README.md"

packages = [
    { include = "dojo", from = "src" },
]

[[tool.poetry.source]]
name = "dojo"
url = "https://gitlab.ics.muni.cz/api/v4/projects/6977/packages/pypi/simple"
priority = "explicit"

[tool.poetry.dependencies]
python = "^3.11"
cyst-core = { version = "*", source = "dojo" }
uvicorn = "^0.23.0"
fastapi = {extras = ["all"], version = "^0.104.0"}
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "^23.3.0"}
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
pytest-mock = "^3.11.1"
aiosqlite = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
